+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/AuthenticationUtils.java
@@ -9,6 +9,8 @@ import io.fabric8.kubernetes.api.model.KeyToPathBuilder;
 import io.fabric8.kubernetes.api.model.Volume;
 import io.fabric8.kubernetes.api.model.VolumeMount;
 import io.strimzi.api.kafka.model.CertSecretSource;
+import io.strimzi.api.kafka.model.KafkaJmxAuthentication;
+import io.strimzi.api.kafka.model.KafkaJmxAuthenticationPassword;
 import io.strimzi.api.kafka.model.authentication.KafkaClientAuthentication;
 import io.strimzi.api.kafka.model.authentication.KafkaClientAuthenticationOAuth;
 import io.strimzi.api.kafka.model.authentication.KafkaClientAuthenticationPlain;
@@ -229,4 +231,21 @@ public class AuthenticationUtils {
             }
         }
     }
+
+    /**
+     * Generates the necessary resources that the Kafka Cluster needs to secure the Kafka Jmx Port
+     *
+     * @param authentication the Authentication Configuration for the Kafka Jmx Port
+     * @param kafkaCluster the current state of the kafka Cluster CR
+     */
+    public static void configureKafkaJmxOptions(KafkaJmxAuthentication authentication, KafkaCluster kafkaCluster)   {
+        if (authentication != null) {
+            if (authentication instanceof KafkaJmxAuthenticationPassword) {
+                kafkaCluster.setJmxAuthenticated(true);
+            }
+        } else {
+            kafkaCluster.setJmxAuthenticated(false);
+        }
+    }
+
 }