+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaAssemblyOperator.java
@@ -281,6 +281,7 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
                 .compose(state -> state.kafkaGenerateCertificates())
                 .compose(state -> state.kafkaAncillaryCm())
                 .compose(state -> state.kafkaBrokersSecret())
+                .compose(state -> state.kafkaJmxSecret())
                 .compose(state -> state.kafkaPodDisruptionBudget())
                 .compose(state -> state.kafkaStatefulSet())
                 .compose(state -> state.kafkaRollingUpdate())
@@ -1728,6 +1729,21 @@ public class KafkaAssemblyOperator extends AbstractAssemblyOperator<KubernetesCl
 
         Future<ReconciliationState> kafkaBrokersSecret() {
             return withVoid(secretOperations.reconcile(namespace, KafkaCluster.brokersSecretName(name), kafkaCluster.generateBrokersSecret()));
+        }
+
+        Future<ReconciliationState> kafkaJmxSecret() {
+            if (kafkaCluster.isJmxAuthenticated()) {
+                Future<Secret> secretFuture = secretOperations.getAsync(namespace, KafkaCluster.jmxSecretName(name));
+                return secretFuture.compose(res -> {
+                    if (res == null) {
+                        return withVoid(secretOperations.reconcile(namespace, KafkaCluster.jmxSecretName(name),
+                                kafkaCluster.generateJmxSecret()));
+                    }
+                    return withVoid(Future.succeededFuture(this));
+                });
+
+            }
+            return withVoid(secretOperations.reconcile(namespace, KafkaCluster.jmxSecretName(name), null));
         }
 
         Future<ReconciliationState> kafkaNetPolicy() {